requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.3.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-dateutil>=2.8.0

